Budgie on single execution
Allows budgie to be fully created by just calling ./budgie.sh,  having an autostart file that executes the setup-desktop.sh without gui interaction required and waiting for a flag file to be created meaning that setup-desktop finished and is safe to start making the budgie layer.

Although it might have a detail or two over there, but the main thing which is having budgie on a single execution is here.