Adding TODO docs
Quick PR to add some TODO docs and don't forget about them in the future. Not sure on when we'll be able to write them, but just to keep track of them.

Probably we need to prioritize the 'Where is huronOS being used' as we're becoming the official OS of the Mexican Olympiad of Informatics.