Making codium able to load .hsm vscode extensions
## Problem
Currently huronOS lacks of support for vs-code extensions as they were not allowed in several competitions because of their built-in features that might significantly offer vantage over other editors.
Since then, more competitions have opened about this, and now the *IOI* offers VScode extensions on their standard image, and it's specified on their environment for the [IOI 2023 virtual machine](https://ioi2023.hu/contest-environment/)

This is a problem since huronOS aims to offer the flexibility of being used for either ICPC-like competitions and IOI-like. At the same time, huronOS doesn't load apps with integrations itself to keep offering flexibilty, this mean that we should not ship vscode already packaged with extensions, but instead offer the extensions as `.hsm` modules that are activatable/deactivatable to be customized by the contest manager which ones to use.

## The IOI approach
The [approach](https://github.com/ioi-2023/contestant-vm/blob/249496007d5387f1acf0875c0bba592f99756112/setup.d/08-install-ides-tools.sh#L55-L66) that the IOI took for setting the extensions on their official environment, is to download the `.vsix` packages and bring them available in a certain directory for the user to manually take it and install it.
We still not knowing why didn't they pre-installed the extension, but the reasoning behind could be related to do not enable the extensions by default and instead allowing the user to install just the extensions they need to.

## My Approach
My approach goes more by auto-enabling all the activated extensions (.hsm) and then let the user to disable them manually on the extensions section of the vscode UI as they're more related to it than with the `.vsix` extension form.
To automatically load extensions, I decided to create an specific directory: `/opt/codium/contestant/extensions` to store them.
This also required to launch vscode with the following command:
```bash
codium --extensions-dir /opt/codium/contestant/extensions
```

There was another issue with this approach, if the user simply typed 'codium' on the terminal to launch vscode, then no extensions would be loaded. To solve this issue, I created a wrapper script lo launch vscode, which will check if the --extensions-dir flag was provided, and in case that not, it will add it.

The final challenge was that when you install vscode extensions, vscode generates a file `/opt/codium/contestant/extensions/extensions.json` which contains all the metadata of the extensions. We couldn't let the extensions `.hsm` files contain this file as that would override betweek each other and end with only the top-most layer being the only one extension loaded.
To solve this issue, we created a directory called `ids/` where each extension saves it's object metadata in a separate file, and then the codium launcher extension create the extensions array that vscode needs on the `extensions.json` file.

> **Note:** Please, note that currently the launcher wrapper will delete the `.obsolete` file that vscode uses to know which extensions are old and have been replaced with newer versions, this is because the current implementation does not cover the vscode auto-updates that override the `extensions.json` file and creates a mismatch with the previously saved `ids/`.

### How to test
- Checkout the top-most PR of this stack to test all the stack together.
- Build vscode `.hsm` layer and all the other extensions
- Launch vscode in solo and check that everything works as usuall
- Run `hmm -u --activate [some_extension_module]` to activate some extension module, close vscode and relaunch it, then verify that now the vscode extension is avialble.
- Test this same behavior with the rest of the stack extensions.