Updates codeblocks default terminal
# Problem

Currently codeblocks opens a new tab in gnome-terminal (if already open) which might turn confusing if already having on the background a gnome-terminal instance and running the code, which might open a new tab on the terminal, but the terminal itself doesn't go to foreground automatically leaving the user believing that the code never executed.

# Solution

This PR fixes that by updating the terminal command by opening a new window every time the user runs the code.