[hos-wallpaper] Simplifies hos-wallpaper
With this change hos-wallpaper assumes it will only apply the wallpaper for the mode specified in the --manual argument, this would simplify things.

Fixes:
This aproach seems to fix an issue where the system powers off in a contest / event state and when turning on and updating to an Always mode, the wallpaper was stuck in a solid blue. It might be possible that there were other transitions that also caused the issue, but if this one is fixed I might think that possible others might be fixed too.