Make 02-firmware system-layer non interactive
## Why is this PR needed?
Currently when building the 02-firmware system layer, several firmware packages require of the acceptance of the license, being an annoying prompt that requires a simple 'ok' to proceed with the automated script.

This blocks the building script and requires human interation.

## What changed?
Now the dbus configuration where the accepted terms are saved, are being pre-filled to avoid the packaged prompting for the acceptance. This enables further automation for the system upper layers build.

## How to test
With a base layer build, run firmware.sh and verify that no prompt appears anymore.