Creating sysforge: A tool for automatically building the huronOS system layers
## Why is this PR needed?
The current build system of the OS requires a lot of human interaction such a frequent reboots and close follow to properly build all the system layers. This impacts the development time and leads to less testing due to the time consuming that system tests take.

## What is this PR changing?
As a first step into build automation, I'm introducing 'sysforge', a tool to persistently create a queue that the utility will check on bootime to execute another script, like the system layer scripts.

This allows the tool to schedule the execution of each of the system layer scripts, one per boot, and this way properly building the system.

Also, this tools is scalable so that it's also possible to build all the software packages (as long as they do not require human interation).

## How to test
The best way to test this tool is by doing a system build.
Save your huronOS-build-tools in a disk drive, make sure to checkout the branch you want to build before launching ´sysforge´.

Build the 1st system-layer and boot from it. Then run `sysforge` terminal and fill the disk, the mount point and the directory of the huronOS-build-tools repository.

Wait for the wizard to finish.