Making sysforge script-agnostic
## Why this PR is needed
The previous PR introduced a *sysforge* implementation that explicitly create each of the scripts for each layer of the system-layers.
However, this implementation can be optimized by reusing the driver-script generation to be executed by the *sysforge-wizard*; this way the function receives a path to a script and generates the appropiate driver-script and enqueue it for build.

## What have changed?
Behavior stills the same, but now the file `system-layers.list` must contain the updated paths (relative to the repository root).

## How to test
Same as previous PR in this stack.